refactor: put auth dto into response package
